refactor: move paths logic to separate file
