+++ packages/zip-it-and-ship-it/src/main.ts
@@ -12,7 +12,8 @@ import { listFunctionsDirectories, resolveFunctionsDirectories } from './utils/f
 import type { ExtendedRoute, Route } from './utils/routes.js'
 
 export { Config, FunctionConfig } from './config.js'
-export { type FunctionsBag, zipFunction, zipFunctions, ZipFunctionOptions, ZipFunctionsOptions } from './zip.js'
+export { zipFunction, zipFunctions, ZipFunctionOptions, ZipFunctionsOptions } from './zip.js'
+export type { FunctionsBag } from './paths.js'
 
 export { ArchiveFormat, ARCHIVE_FORMAT } from './archive.js'
 export type { TrafficRules } from './rate_limit.js'
